get stripe product
