+++ apps/express-server/src/routes/stripe/product/types.ts
@@ -2,5 +2,9 @@ import { type Request } from 'express';
 import type Stripe from 'stripe';
 
 export type CreateProductBody = Stripe.ProductCreateParams;
-
 export type CreateProductRequest = Request<object, object, CreateProductBody>;
+
+export type GetProductBody = {
+	productId: string;
+};
+export type GetProductRequest = Request<GetProductBody>;
