+++ package.json
@@ -31,7 +31,7 @@
     "husky": "^9.0.10",
     "lint-staged": "^15.2.2",
     "prettier": "^3.2.5",
-    "turbo": "^2.4.4"
+    "turbo": "^2.5.0"
   },
   "lint-staged": {
     "*.{json,html,css,sass,scss}": [
